+++ Demo/Keyboard/KeyboardKit+Demo.swift
@@ -1,17 +1,24 @@
 //
-//  AudioFeedback+Demo.swift
-//  KeyboardPro
+//  KeyboardKit+Demo.swift
+//  Keyboard
 //
 //  Created by Daniel Saidi on 2022-02-07.
 //  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
-import KeyboardKitPro
+import KeyboardKit
+
+extension KeyboardAction {
+    
+    static let rocket = character("🚀")
+}
 
 extension Feedback.Audio {
  
-    static let fuse = Self.customUrl(
+    static let rocketFuse = customUrl(
         Bundle.main.url(forResource: "fuse", withExtension: "wav")
     )
+    
+    static let rocketLaunch = customId(1303)
 }