+++ Demo/KeyboardPro/KeyboardViewController.swift
@@ -62,6 +62,16 @@ class KeyboardViewController: KeyboardInputViewController {
         state.keyboardContext.spaceLongPressBehavior = .moveInputCursor
         // state.keyboardContext.spaceLongPressBehavior = .openLocaleContextMenu
         
+        /// 💡 Setup haptic and audio feedback.
+        ///
+        /// The code below enables all haptic feedback, then
+        /// sets up custom audio for the rocket button.
+        let feedback = state.feedbackContext
+        feedback.isHapticFeedbackEnabled = true
+        feedback.audioConfiguration.register(feedback: .rocketFuse, for: .press, on: .rocket)
+        feedback.audioConfiguration.register(feedback: .rocketLaunch, for: .release, on: .rocket)
+        // feedback.isAudioFeedbackEnabled = false
+        
         /// 💡 Setup haptic feedback.
         ///
         /// The code below enables all haptic feedback.
@@ -71,12 +81,11 @@ class KeyboardViewController: KeyboardInputViewController {
         ///
         /// The code below sets up a fuse sound for when the
         /// rocket is pressed, and a system sound on release.
+        let rocket = KeyboardAction.character("🚀")
         state.feedbackContext.audioConfiguration.register(
-            feedback: .fuse, for: .press, on: .character("🚀")
-        )
+            feedback: .rocketFuse, for: .press, on: rocket)
         state.feedbackContext.audioConfiguration.register(
-            feedback: .customId(1303), for: .release, on: .character("🚀")
-        )
+            feedback: .rocketLaunch, for: .release, on: rocket)
         
         /// 💡 Disable autocorrect.
         ///