+++ RELEASE_NOTES.md
@@ -31,6 +31,7 @@ KeyboardKit Pro makes more locales use the iPad Pro layout, and has a new `Keybo
 * `Feedback.Audio` has a new `customUrl` that can play URL-based audio.
 * `Feedback.AudioConfiguration` has new ways to register custom feedback.
 * `Feedback.HapticConfiguration` has new ways to register custom feedback.
+* `FeedbackContext` has new ways to register custom audio/haptic feedback.
 * `KeyboardLayout` has a new `copy()` function that creates a mutable copy.
 * `KeyboardLayout` has new functions for using and tweaking the bottom row.
 * `KeyboardLayout` has a new `inputToolbarLayout()` value builder function.